% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spencer.R
\name{spencer}
\alias{spencer}
\title{Spencer filter}
\usage{
spencer(x, lambda = NULL)
}
\arguments{
\item{x}{A time series (seasonally adjusted).}

\item{lambda}{Box-Cox transformation parameter. If lambda = "auto",
then a transformation is automatically selected using [BoxCox.lambda()].}
}
\value{
A time series after apllying the two-sided Spencer filter.
}
\description{
Filter a series using a 15-term two-sided Spencer filter.
}
\details{
This filter is used in the Bry-Boschan algorithm.
}
